migrated addNewTerm function
